Fix issue where aggregates weren't aggregated and more apis (#17)
* Fix an issue where by only first change is layered on aggregate

* Change DataAccessor to IAggregateHistory

* Test for aggregate layer should work

+++ src/NEvilES.DataStore/EventDb.cs
@@ -4,7 +4,7 @@ namespace NEvilES.DataStore
 {
     public class EventDb
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string Category { get; set; }
         public Guid StreamId { get; set; }
         public Guid TransactionId { get; set; }
@@ -14,7 +14,7 @@ namespace NEvilES.DataStore
         public string Body { get; set; }
         public DateTime When { get; set; }
         public int Version { get; set; }
-        
+
         //Do we really need this?
         public string Metadata { get; set; }
     }
